Use ARF report from the DB
This requires OpenSCAP 1.2.1 and rubygem-openscap 0.4.1

Resolves to OpenSCAP/foreman_openscap#12
