+++ lib/scaptimony/engine.rb
@@ -1,9 +1,5 @@
 module Scaptimony
   class Engine < ::Rails::Engine
     isolate_namespace Scaptimony
-    def self.dir
-      # TODO this should be configurable
-      '/var/lib/foreman/scaptimony'
-    end
   end
 end
